Fixed layout of Read-Only mode of PropertiesWidget.
Fixed populating of combobox elements in PropertiesWidget.
Check solution now appears only when playing, and not on the preview.
Clicking save now makes a call to validate to the puzzle, and if invalid, displays a message.
Added a validation function to Cryptogram that throws an exception when the plaintext is empty or null.
Added testing for the validation function in the appropriate test cases.
Save button now offers some feedback:  If the save is successful, a message is displayed, and the user is returned to the welcome screen.
Removed all author identifiers from presentation layer.
Edited the default constructor for puzzle to have a default name, category, and difficulty.