Reverted User changes. Logic change for adding users.
Now, the idea is that users are not created outside of the manager.
Instead, the manager handles user creation and in turn handles
primaryKey generation.

This behaviour will be unkown to the db and thus the db can still read
and create users when the primary key is known.

I hope that explanation is clear.