Changed edit behaviour
The overall behaviour of save has changed.

When saving a puzzle with ID == DEFAULT_ID, the puzzle gets saved. If
the ID !=DEFAULT_ID then the PuzzleManager tries to find a puzzle with
the same ID and change its values.

So, to edit a puzzle, simply change it's values (except it's ID) and
then save it with the puzzlemanager.